W-who=wh-who&W-when+
Wh-who
For some reason I think this works better. With who, the sound that gets repeated is the wh as opposed to say why, what, when and where  where its the w sound that gets repeated.
However, I'm second guessing myself to be honest. 
I might end up going back and just changing it to who-who, but even that feels strange. Do people even stutter on "who"?
Also fixed more w-w-when/ect